third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/declared/update-expected.txt

This is a testharness.js-based test.
PASS Updating a StylePropertyMap with an unsupported property name throws TypeError
PASS Updating a StylePropertyMap with an null property name throws TypeError
PASS Updating a StylePropertyMap with an invalid CSSStyleValue throws TypeError
PASS Updating a StylePropertyMap with an invalid String throws TypeError
PASS Updating a property with CSSStyleValue updates its value
PASS Updating a list-valued property with CSSStyleValue updates its value
FAIL Updating a custom property with CSSStyleValue updates its value Failed to execute 'update' on 'StylePropertyMap': Invalid propertyName: --foo
PASS Calling StylePropertyMap.update on an empty property model calls update function with null
PASS Calling StylePropertyMap.update on an existing property calls update function with old value
PASS Calling StylePropertyMap.update on an existing list-valued property calls update function with first value
PASS StylePropertyMap.update is case-insensitive
Harness: the test ran to completion.
